Added cost of reoffending example
Why:
We needed not only the iris dataset for basic correctness checks, we also need something which presents the type of challenges which will be faced in the wild.

How does this help:
I've added the cost of reoffending dataset which is highly complex and requires unusual notes/formatting across many different units and categories.