Fixes #49 1, 2 and 4
Why:
There were issues with getting the iris.py to work properly. I believe
that I have corrected most of the issues that were highlighted in the
bug report #49.

What:
I spotted an error in the `GPTable` definition where some code had
been copied and pasted incorrectly.

I spotted some inflexibilities in whether GPWorksheet copied data
when it was working. This means that the nonlocal environment should
be unchanged.

I altered the way that we add a row for column headers. Previously we
had incorrectly used `iloc` when we are now using `loc`.