Fixes #49 1, 2 and 4 (#50)
* Fixes #49 1, 2 and 4

Why:
There were issues with getting the iris.py to work properly. I believe
that I have corrected most of the issues that were highlighted in the
bug report #49.

What:
I spotted an error in the `GPTable` definition where some code had
been copied and pasted incorrectly.

I spotted some inflexibilities in whether GPWorksheet copied data
when it was working. This means that the nonlocal environment should
be unchanged.

I altered the way that we add a row for column headers. Previously we
had incorrectly used `iloc` when we are now using `loc`.

* Delete .~lock.iris_gptable.xlsx#

* Added ~lock to gitignore

Co-authored-by: jazzbuck <40966885+jazzbuck@users.noreply.github.com>
Co-authored-by: David Foster <31405412+foster999@users.noreply.github.com>