Set column width to line length
Previously, auto-width ignored line breaks, meaning it wrapped to the full length of the string. This meant that columns with multi-line strings (e.g. containing the unit at the bottom) were wider than they should be. The updated function calculates the length of the longest line within each string rather than just the full length of the string