Change missing value replacement to error raising (#161)
* Wrap table cells by default

* Set column width to line length

Previously, auto-width ignored line breaks, meaning it wrapped to the full length of the string. This meant that columns with multi-line strings (e.g. containing the unit at the bottom) were wider than they should be. The updated function calculates the length of the longest line within each string rather than just the full length of the string

* Append units to column headings

* update documentation

* Update code examples

* Update tests

* Change missing value replacement to error raising

* Raise error for cells containing only symbols

* Resolve merge conflict

* Raise error for cells containing only symbols

* Remove asyncio/windows_events import

Co-authored-by: Dan Shiloh <daniel.shiloh@ons.gov.uk>